move PrintHandler.js
