+++ index.html
@@ -59,7 +59,7 @@
 
 
 <script type="module">
-    import {printLines} from "./js/bxl/PrintHandler.js";
+    import {printLines} from "./js/PrintHandler.js";
     import getReceiptContent, {keywords} from "./js/receipt.js";
 
     // get arrays of receipt data
